Added Delete method
Added a 'Delete' method to both 'IConfig' interface, and 'Config' class. The
new method deletes a setting both from the file(by reading line by line
and omitting the lines with the settings in them when writing back) and
from the object.

 Readme.md - Added explanation for the new 'Delete' method, and a remark
 about > 2GB config files.

 ConfigParser:
 Config.cs - Added a 'Delete(string setting)' method to the 'IConfig'
interface, and implemented it in the 'Config' class.

 ConfigTester: (Ignore)
 Program.cs - Added debugging for the new 'Delete' method.
 config2.cfg - Changes by debugging